gran salvacion de todo

+++ database/migrations/2018_08_02_212154_create_productos_table.php
@@ -23,9 +23,9 @@ class CreateProductosTable extends Migration
             $table->integer('categorias_id')->unsigned();
             $table->foreign('categorias_id')->references('id')
                     ->on('categorias')
-                    ->onDelete('cascade');            
+                    ->onDelete('cascade');           
             
-            $table->float('precio');
+            $table->double('precio');
 
             $table->timestamps();
         });
